fix CI: extract formatBytes/relativeTime out of mock module
The production knowledge bucket detail page imported these utilities from
``@/lib/mock/cloud``, which the round-2 production-import guard now blocks
at module init. Same code, new home: ``@/lib/format``. The mock module
keeps its guard and stops housing presentation helpers.
